+++ mobile/lib/presentation/widgets/asset_viewer/bottom_sheet.widget.dart
@@ -4,7 +4,6 @@ import 'package:auto_route/auto_route.dart';
 import 'package:collection/collection.dart';
 import 'package:easy_localization/easy_localization.dart';
 import 'package:flutter/material.dart';
-import 'package:flutter/services.dart';
 import 'package:hooks_riverpod/hooks_riverpod.dart';
 import 'package:immich_mobile/constants/enums.dart';
 import 'package:immich_mobile/domain/models/asset/base_asset.model.dart';
@@ -16,8 +15,8 @@ import 'package:immich_mobile/presentation/widgets/album/album_tile.dart';
 import 'package:immich_mobile/presentation/widgets/asset_viewer/asset_viewer.state.dart';
 import 'package:immich_mobile/presentation/widgets/asset_viewer/bottom_sheet/sheet_location_details.widget.dart';
 import 'package:immich_mobile/presentation/widgets/asset_viewer/bottom_sheet/sheet_people_details.widget.dart';
+import 'package:immich_mobile/presentation/widgets/asset_viewer/sheet_tile.widget.dart';
 import 'package:immich_mobile/presentation/widgets/bottom_sheet/base_bottom_sheet.widget.dart';
-import 'package:immich_mobile/providers/haptic_feedback.provider.dart';
 import 'package:immich_mobile/providers/infrastructure/action.provider.dart';
 import 'package:immich_mobile/providers/infrastructure/album.provider.dart';
 import 'package:immich_mobile/providers/infrastructure/asset_viewer/current_asset.provider.dart';
@@ -181,7 +180,7 @@ class _AssetDetailBottomSheet extends ConsumerWidget {
           spacing: 12,
           children: [
             if (albums.isNotEmpty)
-              _SheetTile(
+              SheetTile(
                 title: 'appears_in'.t(context: context).toUpperCase(),
                 titleStyle: context.textTheme.labelMedium?.copyWith(
                   color: context.textTheme.labelMedium?.color?.withAlpha(200),
@@ -233,7 +232,7 @@ class _AssetDetailBottomSheet extends ConsumerWidget {
           future: assetMediaRepository.getOriginalFilename(asset.id),
           builder: (context, snapshot) {
             final displayName = snapshot.data ?? asset.name;
-            return _SheetTile(
+            return SheetTile(
               title: displayName,
               titleStyle: context.textTheme.labelLarge,
               leading: Icon(
@@ -250,7 +249,7 @@ class _AssetDetailBottomSheet extends ConsumerWidget {
         );
       } else {
         // For remote assets, use the name directly
-        return _SheetTile(
+        return SheetTile(
           title: asset.name,
           titleStyle: context.textTheme.labelLarge,
           leading: Icon(
@@ -269,7 +268,7 @@ class _AssetDetailBottomSheet extends ConsumerWidget {
     return SliverList.list(
       children: [
         // Asset Date and Time
-        _SheetTile(
+        SheetTile(
           title: _getDateTime(context, asset),
           titleStyle: context.textTheme.bodyMedium?.copyWith(fontWeight: FontWeight.w600),
           trailing: asset.hasRemote && isOwner ? const Icon(Icons.edit, size: 18) : null,
@@ -279,7 +278,7 @@ class _AssetDetailBottomSheet extends ConsumerWidget {
         const SheetPeopleDetails(),
         const SheetLocationDetails(),
         // Details header
-        _SheetTile(
+        SheetTile(
           title: 'exif_bottom_sheet_details'.t(context: context),
           titleStyle: context.textTheme.labelMedium?.copyWith(
             color: context.textTheme.labelMedium?.color?.withAlpha(200),
@@ -290,7 +289,7 @@ class _AssetDetailBottomSheet extends ConsumerWidget {
         buildFileInfoTile(),
         // Camera info
         if (cameraTitle != null)
-          _SheetTile(
+          SheetTile(
             title: cameraTitle,
             titleStyle: context.textTheme.labelLarge,
             leading: Icon(Icons.camera_alt_outlined, size: 24, color: context.textTheme.labelLarge?.color),
@@ -301,7 +300,7 @@ class _AssetDetailBottomSheet extends ConsumerWidget {
           ),
         // Lens info
         if (lensTitle != null)
-          _SheetTile(
+          SheetTile(
             title: lensTitle,
             titleStyle: context.textTheme.labelLarge,
             leading: Icon(Icons.camera_outlined, size: 24, color: context.textTheme.labelLarge?.color),
@@ -315,77 +314,6 @@ class _AssetDetailBottomSheet extends ConsumerWidget {
         // padding at the bottom to avoid cut-off
         const SizedBox(height: 100),
       ],
-    );
-  }
-}
-
-class _SheetTile extends ConsumerWidget {
-  final String title;
-  final Widget? leading;
-  final Widget? trailing;
-  final String? subtitle;
-  final TextStyle? titleStyle;
-  final TextStyle? subtitleStyle;
-  final VoidCallback? onTap;
-
-  const _SheetTile({
-    required this.title,
-    this.titleStyle,
-    this.leading,
-    this.subtitle,
-    this.subtitleStyle,
-    this.trailing,
-    this.onTap,
-  });
-
-  void copyTitle(BuildContext context, WidgetRef ref) {
-    Clipboard.setData(ClipboardData(text: title));
-    ImmichToast.show(
-      context: context,
-      msg: 'copied_to_clipboard'.t(context: context),
-      toastType: ToastType.info,
-    );
-    ref.read(hapticFeedbackProvider.notifier).selectionClick();
-  }
-
-  @override
-  Widget build(BuildContext context, WidgetRef ref) {
-    final Widget titleWidget;
-    if (leading == null) {
-      titleWidget = LimitedBox(
-        maxWidth: double.infinity,
-        child: Text(title, style: titleStyle),
-      );
-    } else {
-      titleWidget = Container(
-        width: double.infinity,
-        padding: const EdgeInsets.only(left: 15),
-        child: Text(title, style: titleStyle),
-      );
-    }
-
-    final Widget? subtitleWidget;
-    if (leading == null && subtitle != null) {
-      subtitleWidget = Text(subtitle!, style: subtitleStyle);
-    } else if (leading != null && subtitle != null) {
-      subtitleWidget = Padding(
-        padding: const EdgeInsets.only(left: 15),
-        child: Text(subtitle!, style: subtitleStyle),
-      );
-    } else {
-      subtitleWidget = null;
-    }
-
-    return ListTile(
-      dense: true,
-      visualDensity: VisualDensity.compact,
-      title: GestureDetector(onLongPress: () => copyTitle(context, ref), child: titleWidget),
-      titleAlignment: ListTileTitleAlignment.center,
-      leading: leading,
-      trailing: trailing,
-      contentPadding: leading == null ? null : const EdgeInsets.only(left: 25),
-      subtitle: subtitleWidget,
-      onTap: onTap,
     );
   }
 }
